# oj-forntend/src/api/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "oj-backend",
    "version": "1.0",
    "title": "接口文档"
  },
  "host": "localhost:8101",
...
  "paths": {
...
    "/api/post/add": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "addPost",
        "operationId": "addPostUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postAddRequest",
            "description": "postAddRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostAddRequest",
              "originalRef": "PostAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«long»",
              "originalRef": "BaseResponse«long»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/delete": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "deletePost",
        "operationId": "deletePostUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "deleteRequest",
            "description": "deleteRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteRequest",
              "originalRef": "DeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«boolean»",
              "originalRef": "BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/edit": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "editPost",
        "operationId": "editPostUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postEditRequest",
            "description": "postEditRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostEditRequest",
              "originalRef": "PostEditRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«boolean»",
              "originalRef": "BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/get/vo": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "summary": "getPostVOById",
        "operationId": "getPostVOByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«PostVO»",
              "originalRef": "BaseResponse«PostVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/list/page/vo": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "listPostVOByPage",
        "operationId": "listPostVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostQueryRequest",
              "originalRef": "PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»",
              "originalRef": "BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/my/list/page/vo": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "listMyPostVOByPage",
        "operationId": "listMyPostVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostQueryRequest",
              "originalRef": "PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»",
              "originalRef": "BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/search/page/vo": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "searchPostVOByPage",
        "operationId": "searchPostVOByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostQueryRequest",
              "originalRef": "PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»",
              "originalRef": "BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/post/update": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "updatePost",
        "operationId": "updatePostUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postUpdateRequest",
            "description": "postUpdateRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostUpdateRequest",
              "originalRef": "PostUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«boolean»",
              "originalRef": "BaseResponse«boolean»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/user/get/login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getLoginUser",
        "operationId": "getLoginUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponse«LoginUserVO»",
              "originalRef": "BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BaseResponse«LoginUserVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/LoginUserVO",
          "originalRef": "LoginUserVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«LoginUserVO»"
    },
    "BaseResponse«Page«PostVO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page«PostVO»",
          "originalRef": "Page«PostVO»"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Page«PostVO»»"
    },
...
    "BaseResponse«PostVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/PostVO",
          "originalRef": "PostVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«PostVO»"
    },
...
    "BaseResponse«boolean»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«boolean»"
    },
...
    "BaseResponse«long»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«long»"
    },
...
    "DeleteRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "DeleteRequest"
    },
    "LoginUserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "LoginUserVO"
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean"
        },
        "column": {
          "type": "string"
        }
      },
      "title": "OrderItem"
    },
    "Page«PostVO»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PostVO",
            "originalRef": "PostVO"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«PostVO»"
    },
...
    "PostAddRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      },
      "title": "PostAddRequest"
    },
    "PostEditRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      },
      "title": "PostEditRequest"
    },
...
    "PostQueryRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "favourUserId": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "notId": {
          "type": "integer",
          "format": "int64"
        },
        "orTags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pageSize": {
          "type": "integer",
          "format": "int64"
        },
        "searchText": {
          "type": "string"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostQueryRequest"
    },
...
    "PostUpdateRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        }
      },
      "title": "PostUpdateRequest"
    },
    "PostVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "hasFavour": {
          "type": "boolean"
        },
        "hasThumb": {
          "type": "boolean"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tagList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "user": {
          "$ref": "#/definitions/UserVO",
          "originalRef": "UserVO"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostVO"
    },
...
    "UserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserVO"
    }
  }
}